chore(memory): AES 제거 작업 잔재 정리
- migrations.rs: memory_secret 테이블 주석에서 "AES-256-GCM 암호화" 표현
  제거 → "평문 저장 + IPC owner 격리" 로 정정
- lib.rs: 평문화 후 남아 있던 i64→i64 불필요 cast 제거 (clippy)
- docs/index.md: IPC 메서드 인벤토리에 memory.secret.{put,get,delete,
  list,exists,count,scopes,stats} 추가 (이전부터 누락)

+++ crates/tasty-memory/src/lib.rs
@@ -299,7 +299,7 @@ impl MemoryStore {
             .optional()?;
 
         // Quota: 추가될 byte 가 regular total 한도를 넘기지 않는지.
-        let existing_size = existing.as_ref().map(|(_, _, sz)| *sz as i64).unwrap_or(0);
+        let existing_size = existing.as_ref().map(|(_, _, sz)| *sz).unwrap_or(0);
         let current_used: i64 = tx.query_row(
             "SELECT COALESCE(SUM(LENGTH(value)), 0) FROM memory",
             [],
